requirements.txt
azure-ai-documentintelligence
openai
python-dotenv
requests
azure-core
PyMuPDF
pydantic
fastapi
pyyaml
jinja2
azure-ai-projects
azure-identity
openpyxl
pandas
comtypes
pywin32
python-pptx
python-docx
reportlab
httpx
pytest
python-multipart